requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
pydantic==2.5.0
email-validator==2.1.0
